+++ lib/src/cli/cli.dart
@@ -55,7 +55,12 @@ class _Cmd {
   }
 }
 
+final _ignoredDirectories = RegExp(
+  'ios|android|windows|linux|macos|.symlinks|.plugin_symlinks|.dart_tool|build',
+);
+
 bool _isPubspec(FileSystemEntity entity) {
+  if (entity.path.contains(_ignoredDirectories)) return false;
   if (entity is! File) return false;
   return p.basename(entity.path) == 'pubspec.yaml';
 }