+++ lib/src/commands/commands.dart
@@ -1,1 +1,2 @@
 export 'create.dart';
+export 'packages.dart';