+++ lib/src/templates/dart_package/dart_package.dart
@@ -1,0 +1,2 @@
+export 'dart_package_bundle.dart';
+export 'dart_package_template.dart';